pyproject.toml
[tool.poetry]
name = "mmcore"
version = "0.5.2"
description = "mmcore"
authors = ["Andrew Astakhov <aa@contextmachine.ru>", ]
license = "Apache License Version 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
graphene = { version = "^3.0.0" }
pydantic = { version = "1.10.7" }
starlette = "0.22.0"
ipython = { version = "^8.14.0" }
termcolor = { version = "^2.3.0" }
setuptools = { version = "^67.8.0" }
fastapi = { version = "*", extras = ["all"] }
earcut = { version = "^1.1.5" }
numpy = { version = "^1.24.1" }
shapely = "^2.0.1"
shell = "^1.0.1"
compas = "^1.17.5"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
redis = "^4.6.0"
requests = "^2.31.0"
scipy = "^1.11.1"
dill = "^0.3.7"
httpx = "^0.24.1"
rpyc = "^5.3.1"
typing-extensions = "^4.7.1"
jinja2 = "^3.1.2"
colored = "^2.2.3"
boto3 = "^1.28.16"
strawberry-graphql = { extras = ["fastapi"], version = "^0.198.0" }
ujson = "^5.8.0"
pyquaternion = "^0.9.9"
multiprocess = "^0.70.15"
cloudpickle = "^2.2.1"
more-itertools = "~10.1.0"
uvicorn = { version = "^0.23.2", extras = ["standard"] }
randomname = "^0.2.1"
geomdl = "^5.3.1"
rich = "^13.5.2"
docker = "^6.1.3"
pyvis = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
